requirements.txt
python-binance==1.0.29
python-dotenv==1.1.1
pandas==2.3.3
numpy==2.2.6
pandas-ta==0.4.71b0
experta @ git+https://github.com/nilp0inter/experta.git@develop
scikit-learn>=1.3.0
joblib>=1.3.0
